requirements.txt
pyspark
apache-airflow
kafka-python
great_expectations
boto3
